requirements.txt
google-genai==1.49.0
streamlit==1.51.0
pandas==2.3.3
python-dotenv==1.2.1
google-adk==1.18.0